Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15501 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/ADSrc/ADDriver.h
@@ -187,6 +187,6 @@ protected:
     int ADStringFromServer; 
     #define LAST_AD_PARAM ADStringFromServer   
 };
-#define NUM_AD_PARAMS (&LAST_AD_PARAM - &FIRST_AD_PARAM + 1)
+#define NUM_AD_PARAMS ((int)(&LAST_AD_PARAM - &FIRST_AD_PARAM + 1))
 
 #endif
